--- a/src/components/RenderForm.js
+++ b/src/components/RenderForm.js
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { createRef } from 'react'
 import PropTypes from 'prop-types'
 import Form from '@rjsf/core'
 import 'bootstrap/dist/css/bootstrap.min.css'
@@ -6,6 +6,7 @@ import { customizeValidator } from '@rjsf/validator-ajv8'
 import { widgets } from '../custom-ui/CustomWidgets'
 import { uiSchema } from '../custom-ui/CustomUISchema'
 import { AJV_OPTIONS } from '../utilities/schemaHandlers'
+import { deepEquals } from '@rjsf/utils'
 
 function RenderForm (props) {
   /*
@@ -18,7 +19,36 @@ function RenderForm (props) {
     Form object
   */
   const { schemaType, schema, formData } = props
+  const formRef = createRef()
   const validator = customizeValidator(AJV_OPTIONS)
+
+  /**
+   * Custom onBlur event handler to omit extra data from formData.
+   * Based on omitExtraData logic in RJSF Form component.
+   */
+  const omitExtraDataOnBlur = () => {
+    // NOTE: There is a bug in RJSF `omitExtraData` logic causing validation errors.
+    // This is a workaround to omit extra data whenever a user leaves an input field.
+    // We avoid using `liveOmit` prop due to perf issues.
+
+    // TODO: remove this function once bug is fixed in RJSF.
+    const {
+      schemaUtils: _schemaUtils,
+      schema: _schema,
+      formData: _formData
+    } = formRef.current.state
+    const retrievedSchema = _schemaUtils.retrieveSchema(_schema, _formData)
+    const pathSchema = _schemaUtils.toPathSchema(retrievedSchema, '', _formData)
+    const fieldNames = formRef.current.getFieldNames(pathSchema, _formData)
+    const newFormData = formRef.current.getUsedFormData(_formData, fieldNames)
+
+    if (!deepEquals(_formData, newFormData)) {
+      formRef.current.setState({
+        ...formRef.current.state,
+        formData: newFormData
+      })
+    }
+  }
 
   async function saveFilePicker (event) {
     /*
@@ -43,12 +73,15 @@ function RenderForm (props) {
 
   if (schema) {
     return (
-      schema && <Form schema={schema}
+      schema && <Form
+        ref={formRef}
+        schema={schema}
         formData={formData}
         validator={validator}
         uiSchema={uiSchema}
         widgets={widgets}
         onSubmit={saveFilePicker}
+        onBlur={omitExtraDataOnBlur}
         omitExtraData
         noHtml5Validate >
       </Form>