fixes #162 
- Removed `liveOmit` RJSF prop due to performance issues
- Added custom `omitExtraDataOnBlur` event handler as workaround for RJSF  `omitExtraData` bug
- Added unit tests for RenderForm component, minor refactor to move `saveToJSONFile` to utils.

Before: keypress event takes ~200ms for empty Rig and >2s for complex "filled" Rig
Now: keypress event takes <50ms for empty Rig and <300ms for complex "filled" Rig